lesson08/django_blog_lesson08/mysite/router.py
from rest_framework import routers
from api.viewsets import BloggingViewSet

router = routers.DefaultRouter()
router.register('blogging', BloggingViewSet, base_name='blogging')

for post in router.urls:
    print(post, '\n')
