lesson08/django_blog_lesson08/requirements.txt
Django==2.2.3
pytz==2018.9
django-allauth==0.39.1
djangorestframework==3.10.2
